+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 0.2.4 (WIP)
 
 * Update dependencies
+* Added clamp value toggle as a workaround for [#12](https://github.com/iliags/click_storm/issues/12)
 
 ## 0.2.3